WIP on pytest suite, add more linting tools, update deps
