+++ .pre-commit-config.yaml
@@ -1,17 +1,32 @@
 # pre-commit run --all-files
 repos:
--   repo: https://github.com/pre-commit/pre-commit-hooks
+  - repo: https://github.com/pre-commit/pre-commit-hooks
     rev: v3.4.0
     hooks:
-    -   id: end-of-file-fixer
-    -   id: trailing-whitespace
-    -   id: mixed-line-ending
--   repo: https://github.com/psf/black
-    rev: 20.8b1
+      - id: check-ast
+      - id: check-json
+      - id: check-merge-conflict
+      - id: check-toml
+      - id: check-yaml
+      - id: debug-statements
+      - id: detect-aws-credentials
+        args: ["--allow-missing-credentials"]
+      - id: detect-private-key
+      - id: end-of-file-fixer
+      - id: mixed-line-ending
+      - id: no-commit-to-branch
+      - id: trailing-whitespace
+  - repo: https://github.com/psf/black
+    rev: 21.6b0
     hooks:
-    -   id: black
--   repo: https://github.com/asottile/blacken-docs
-    rev: v1.9.1
+      - id: black
+  - repo: https://github.com/asottile/blacken-docs
+    rev: v1.10.0
     hooks:
-    -   id: blacken-docs
-        additional_dependencies: [black==20.8b1]
+      - id: blacken-docs
+        additional_dependencies: [black==21.6b0]
+  - repo: https://github.com/flakehell/flakehell
+    rev: v.0.8.0
+    hooks:
+      - name: Run flakehell static analysis tool
+        id: flakehell
